+++ src/main/java/org/mbari/mxm/provider_client/responses/MissionStatusResponse.java
@@ -6,8 +6,11 @@ public record MissionStatusResponse(PMResponse result) {
 
   @Data
   public static class PMResponse {
-    // captured as `providerMissionId` in MXM mission model
     public String missionId;
     public String status;
+
+    public String missionTplId;
+
+    public String providerMissionId;
   }
 }